# docs/swagger.yml
swagger: "2.0"
info:
  title: "REST API Documentatie WP3"
  description: >
    In onze REST API maken we gebruik van JWT voor authenticatie. Alle beschermde routes vereisen een "access_token". Deze wordt gegenereerd bij elke succesvolle inlogpoging.
  version: "2.0"
...
paths:
...
  /api/organizations:
    get:
      tags:
        - "Organizations"
      summary: "Geeft een lijst van alle organisaties"
      description: "Geeft een lijst van alle organisaties met betrekking tot de optionele zoekterm en rol van degene die de request maakt."
      parameters:
        - name: zoekterm
          in: query
          type: string
          required: false
          description: Pas de weergave van organisaties aan op basis van deze invoer.
        - name: sorteer_op
          in: query
          type: string
          required: false
          description: Sorteer de weergave van organisaties op basis van deze invoer.
        - name: volgorde
          in: query
          type: string
          required: false
          description: Op- of aflopend
      responses:
        200:
          description: "Alle organisaties opgehaald"
          schema:
            type: "object"
            properties:
              current_page:
                type: integer
                example: 1
              data:
                type: "array"
                items:
                  $ref: "#/definitions/GetOrganization"
              message:
                type: "string"
                example: "string"
              success:
                type: "boolean"
                example: true
              total_pages:
                type: integer
                example: 1
              total_records:
                type: integer
                example: 1
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Geen toegang tot deze informatie"
            
      security:
        - jwt_auth: []

    post:
      tags:
        - "Organizations"
      summary: " Verwerkt de registratiepoging van een organisatie."
      description: "Bij een succesvolle registratie worden de gegevens opgeslagen in de database."
      parameters:
        - name: "body"
          in: "body"
          description: |
            **Invoervelden**
            name, type, website, description, contact, email, password, number, details
          required: true
          schema:
            $ref: "#/definitions/CreateOrganization"    
      responses:
        201:
          description: "Registratie van de organisatie is succesvol opgeslagen"
          schema:
            $ref: "#/definitions/ApiResponseInteger"
        500:
          description: "Registratie mislukt"
  /api/organizations/{organization_id}:
    get:
      tags:
        - "Organizations"
      summary: "Geeft alle gegevens van de organisatie met de opgegeven organization_id"
      description: "Haalt de gegevens op van de organisatie aan de hand van de organization_id"   
      parameters:
      - name: "organization_id"
        in: "path"
        description: De unieke ID van de organisatie
        required: true
        type: "string"
      responses:
        200:
          description: "Organisatiegegevens succesvol opgehaald"
          schema:
            type: "object"
            properties:
              data:
                type: "object"
                example: "#/definitions/GetOrganisation"
              message:
                type: "string"
                example: "string"
              success:
                type: "boolean"
                example: true
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Geen toegang tot deze informatie"
        404:
          description: "Organisatie niet gevonden"
      security:
        - jwt_auth: []

    put:
      tags:
        - "Organizations"
      summary: "Past gegevens aan van de organisatie met de opgegeven organization_id"
      description: "Past gegevens aan van de organisatie met de opgegeven organization_id."
      parameters:
      - name: "organization_id"
        in: "path"
        description: De unieke ID van de organisatie
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Gegevens om aan te passen met de nieuwe waarden"
        required: true
        schema:
          $ref: "#/definitions/CreateOrganization"
      responses:
        200:
          description: "Organisatiegegevens succesvol aangepast"
          schema:
            $ref: "#/definitions/ApiResponseInteger"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere organisatie aan te passen"
      security:
        - jwt_auth: []

    patch:
      tags:
        - "Organizations"
      summary: "Gebruik om alleen de status van de organisatie met de opgegeven organization_id aan te passen"
      description: "Gebruik om alleen de status van de organisatie met de opgegeven organization_id aan te passen"
      parameters:
      - name: "organization_id"
        in: "path"
        description: De unieke ID van de organisatie
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Alleen de status van een organisatie"
        required: true
        schema:
          type: "object"
          required:
              - status
          properties:
            status:
              type: "string"
              example: "string"
              enum: ["Nieuw", "Goedgekeurd", "Afgekeurd"]
              description: "Status van een organisatie"
      responses:
        200:
          description: "Status van de organisatie succesvol aangepast"
          schema:
            $ref: "#/definitions/ApiResponseInteger"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere organisatie aan te passen"
      security:
        - jwt_auth: []

    delete:
      tags:
        - "Organizations"
      summary: "Verwijdert alle gegevens van de organisatie met de opgegeven organization_id"
      description: "Verwijdert alle gegevens van de organisatie met de opgegeven organization_id."
      parameters:
      - name: "organization_id"
        in: "path"
        description: De unieke ID van de organisatie
        required: true
        type: "string"
      responses:
        200:
          description: "Organisatie succesvol verwijderd"
          schema:
            $ref: "#/definitions/ApiResponseInteger"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere organisatie te verwijderen"
      security:
        - jwt_auth: []
...
  /api/subscriptions:
    get:
      tags:
        - "Subscriptions"
      summary: "Geeft een lijst van alle onderzoekensaanmeldingen"
      description: "Geeft een lijst van alle onderzoekensaanmeldingen met betrekking tot de optionele zoekterm en rol van degene die een request maakt."
      parameters:
        - name: zoekterm
          in: query
          type: string
          required: false
          description: Pas de weergave van onderzoekensaanmeldingen aan op basis van deze invoer.
        - name: sorteer_op
          in: query
          type: string
          required: false
          description: Sorteer de weergave van onderzoekensaanmeldingen op basis van deze invoer.
        - name: volgorde
          in: query
          type: string
          required: false
          description: Op- of aflopend
      responses:
        200:
          description: "Alle onderzoekensaanmeldingen opgehaald"
          schema:
            type: "object"
            properties:
              current_page:
                type: integer
                example: 1
              data:
                type: "array"
                items:
                  $ref: "#/definitions/GetSubscription"
              message:
                type: "string"
                example: "string"
              success:
                type: "boolean"
                example: true
              total_pages:
                type: integer
                example: 1
              total_records:
                type: integer
                example: 1
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Geen toegang tot deze informatie"
      security:
        - jwt_auth: []

    post:
      tags:
        - "Subscriptions"
      summary: "Verwerkt de onderzoeksaanmelding van een gebruiker"
      description: "Bij een succesvolle onderzoeksaanmelding wordt deze opgeslagen in de database."
      parameters:
      - name: "research_id"
        in: "path"
        description: "De unieke ID van het onderzoek"
        required: true
        type: "string"
      responses:
        201:
          description: "Onderzoeksaanmelding succesvol opgeslagen"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "U kunt geen onderzoeksaanmelding doen"
      security:
        - jwt_auth: []
  /api/subscriptions/{research_id}:
    delete:
      tags:
        - "Subscriptions"
      summary: "Verwerkt de onderzoeksafmelding van een gebruiker"
      description: "Verwerkt de onderzoeksafmelding van een gebruiker"
      parameters:
      - name: "research_id"
        in: "path"
        description: "De unieke ID van het onderzoek"
        required: true
        type: "string"
      responses:
        200:
          description: "Onderzoeksafmelding succesvol verwerkt"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om deze onderzoeksafmelding te verwijderen"
      security:
        - jwt_auth: []
securityDefinitions:
  jwt_auth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Voer in: Bearer <access_token>

      Voorbeeld: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
definitions:
  ApiResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
        example: true
      message:
        type: "string"
        example: "string"
...
  ApiResponseInteger:
    type: "object"
    properties:
      success:
        type: "boolean"
        example: true
      message:
        type: "string"
        example: "string"
      data:
        type: "integer"
        example: 0
...
  CreateOrganization:
    type: "object"
    properties:
      contact:
        type: "string"
        example: "string"
      description:
        type: "string"
        example: "string"
      details:
        type: "string"
        example: "string"
      email:
        type: "string"
        example: "string"
      name:
        type: "string"
        example: "string"
      number:
        type: "string"
        example: "string"
      password:
        type: "string"
        example: "string"
      type:
        type: "string"
        example: "string"
      website:
        type: "string"
        example: "string"
    xml:
      name: "CreateOrganization"
  GetOrganization:
    type: "object"
    properties:
      beschrijving:
        type: "string"
        example: "string"
      contactpersoon:
        type: "string"
        example: "string"
      email:
        type: "string"
        example: "string"
      organisatie_id:
        type: "string"
        example: "string"
      organisatie_naam:
        type: "string"
        example: "string"
      organisatie_type:
        type: "string"
        example: "string"
      overige_details:
        type: "string"
        example: "string"
      status:
        type: "string"
        example: "string"
      telefoonnummer:
        type: "string"
        example: "string"
      wachtwoord:
        type: "string"
        example: "string"
      website:
        type: "string"
        example: "string"
    xml:
      name: "GetOrganization"
...
  GetSubscription:
    type: "object"
    properties:
      begindatum:
        type: "string"
        example: "string"
        format: date
      einddatum:
        type: "string"
        example: "string"
        format: date
      onderzoek_id:
        type: "integer"
        example: 0
      onderzoek_type:
        type: "string"
        example: "string"
      status:
        type: "string"
        example: "string"
      titel:
        type: "string"
        example: "string"
    xml:
      name: "GetSubscription"